backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
yfinance>=0.2.36
feedparser>=6.0.0
cachetools>=5.3.0
httpx>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
kiteconnect>=5.0.0
pandas>=2.0.0
python-multipart>=0.0.9
